Statistics binning range can be set manually
Until now, there was no way to manually define the range of the binning
functionality in the Statistics program. For example, if you wanted a
histogram ranged 0 to 100, but your data had a maximum value of 79, then
the last bin would always finish at 79.

To address this issue, a `--manualbinrange' option has been added. When
called, the bin organizing function will use the values given to
`--greaterequal' or `--lessthan' to set the range manually.