Added nice quote from Nature to Science and its tools
Today I came across this nice article by many well-known statisticians that
was published in Nature's November 28th issue: "Five ways to fix
statistics" (https://www.nature.com/articles/d41586-017-07522-z). It was so
similar to what Anscombe was arguing in 1973, so I thought it fits well and
further solidifes the discussion of the "Science and its tools" section of
the book and added it.