Match output names keep directory of given output
Until now, when Match was asked to make two files with a given output
filename, it will ignore the directory of the output.

This is because it uses the standard `gal_checkset_automatic_output'
function which is designed for making an output filenames from THE INPUT
filename. Therefore, as described in Automatic output, it will ignore the
directory of the input's name by default.

However, Match can have two output file names (when the output is requested
to be a text file). In such cases, it also uses
`gal_checkset_automatic_output' to generate the (new) output file names
from the single output name given to it, which caused the problem above.

To fix the issue, when the output file name is used to define the new
output names, Match now temporarily sets `p->cp.keepinputdir' to 1.

This fixes bug #53226.