Statistics program resets blank flags after applying limits
After applying the limits, the Statistics program wouldn't reset the blank
flags of the dataset. Even though the out-of-range fields of the dataset
are set to blank. This wasn't visible in most tests until now because they
were mostly on floating point types were the blank value is NaN is wouldn't
cause any problem in minimum and maximum checks.

Because of this, the `CHECK_BLANK' argument to `GAL_TILE_PARSE_OPERATE' in
`gal_statistics_minimum' and `gal_statistics_maximum' were also set to 1.

This fixes bug #53230.