NoiseChisel checks if any growth is possible
Until now, when doing the quantile growth, NoiseChisel wouldn't actually
check if any pixels existed for growth or not. So when the input data and
parameters created conditions where no growth was possible (no pixels
existed to grow), NoiseChisel would crash. A check has been added, so in
such conditions, NoiseChisel acts as if no growth was requested (the
verbose output also explicity says that no growth happened).

This bug was found during a test run of NoiseChisel on data provided by
Nima Dehdilani and Sara Yousefi.

This fixes bug #53268.