Input kernel no longer mandatory for NoiseChisel
NoiseChisel no longer needs an input kernel to operate. In such situations,
from now on, a value of `none' can be passed to the `--kernel' option.

During the work, I noticed that the `INFO_NFF' macro defined in
`bin/noisechisel/clumps.c' doesn't do what its name suggests (count
numbers). It actually keeps the sum of pixels above zero, not the number
(as explained in the comments). So the name and comments of this variable
were also corrected.

This issue (to disable convolution), was motivated by a test run and
discussion with Aaron Watkins.