NoiseChisel checks expantion when removing false detections
Until now, false detections would be removed in the labeled image only when
no expansion was necessary. However, as described in commit 756b312
(NoiseChisel checks if any growth is possible) some conditions (objects
with sharp edges) were found that even though the user has requested
growth, no pixels satisfy the growth condition. Therefore the labeled image
wouldn't be over-written and NoiseChisel would crash if there were any
false detections (which remained in the labeled image).

To fix this problem, the `detection_remove_false_initial' function which is
in charge of re-labeling (removing) the false detections, now also counts
how many pixels are necessary for the growth step and re-labels the labeled
image if necessary.

This is a continuation of the fix of bug #53268.