NoiseChisel resetting labeled image in no growth
When there is no growth, the previous commit still didn't reset the labels
(it was actually corrected after a failed `make check' and the original
problem was never re-checked after that!). So a check is now added to
specifically parse over the array and re-set the labels (remove false
detections) if there is no expansion.