Range corrected for makeCatalog's clump tiles for upperlimit
When a clump image is requested along with the upper limit measurements,
MakeCatalog creates a "tile" for the clump to enable easy and fast parsing
of the random positions. Until now, the area parsed for each clump when
defining the tile was actually the whole image, not the detected object
(due to a confusion between the actual image size and the object's tile
size).

Since the object label was checked, this didn't actually cause a bad
measurement, but it could caused a crash when the object was close to the
image edge. With this commit, the issue is solved.

This fixes bug #53295.