Position in upperlimit distribution (sigma and quantile)
Three new options are now added to MakeCatalog: two will allow identifying
the position of an object or clump within the upper-limit distribution, and
another can be used to identify the upper limit distribution irrespective
of the `--upnsigma' value.

Also, while working on the quantile position, I noticed that the fact that
`gal_statistics_quantile_function' only returned a NaN when the value was
out of range was not too meaningful (is it smaller than the minimum or
larger than the maximum for example). So now, it returns `inf' or `-inf' to
specify in which direction the value is out of range.