NoiseChisel checks if there was any detection
Until now, NoiseChisel assumed that there is always a detection. It would
thus crash if there wasn't any. So a check was added after the detection
step so NoiseChisel continues only when there are any detections.

This issue came up in a dataset provided by Sara Yousefi Taemeh.

This fixes bug #53304.