Fits program opening file fixed for keyword editing
When any of the commands to edit FITS keywords (except `--delete') were
given to the Fits program, it would crash with a segmentation fault!

The problem was that the call to CFITSIO to read in the FITS file
information was only being read for the `--delete' option and not the rest
of the operations. This issue is fixed now.

I also noticed that when a CFITSIO error came up (for example a keyword
that has been asked to be deleted doesn't exist), we wouldn't reset the
`status' value. So none of the other operations would be done and in the
end CFITSIO would complain (which is not the expected operation in Fits
when `--quitonerror' is not called).

The problem was found by Johannes Zabl.