TIFF files can be used as input
With this commit, Gnuastro contains a library wrapper for the `libtiff'
library to read TIFF data into memory for use within Gnuastro. As a result,
ConvertType can now also read TIFF files. Currently writing of TIFF files
is not yet supported.

Also, I found out that `gal_list_data_add' was not returning the proper
array! It was returning the last node of the `newnode' list, not the
first/full list.

A re-run of Autoconf is necessary after pulling this commit.