gal_statistics_number returns size_t dataset
Until now, `gal_statistics_number' would return a dataset with type
`uint64_t', but that is a little hard to manage (especially for 32-bit
systems). Also, in Gnuastro, we follow the convention of using `size_t' for
all countings and sizes. So the output dataset of this function now has a
`size_t' type (dependent on the host system).

Also, Lee Kelvin kindly noted that NoiseChisel's error message when enough
tiles can't be found for interpolation of the quantile threshold value is
not too clear. Indeed, since the error would be printed by multiple
threads, it was hard to read and very dense. So a test is now added (thanks
to the `gal_statistics_number' function above) to see how many tiles passed
and report an error before NoiseChisel goes into
interpolation/mult-threaded mode. Since several other users had previously
also complained about not understanding the problem, a very thorough
explanation was added to the error message along with suggestions and links
for further reading.