Added astbuildprog.conf to the files to clean
BuildProgram's configuration file is not a source file, its a built file
(since we need to possibly add header and library directories to it during
the build). However, no `CLEANFILES' variable was defined in
`bin/buildprog/Makefile.am'. As a result, it wouldn't get cleaned. This
variable has now been added so the configuration file will be cleaned from
now on.

This problem was reported by Johannes Zabl.