Correceted writing of library and include dirs for Buildprogram
When we were writing the user-given library and include directories into
the BuildProgram configuration file, with the `x"$$i" != x', we were also
checking if the value within the user's given values wasn't empty! For some
reason, this causes Mac OS systems to not write the file at all. Even when
there was no trouble, this appears to also be redundant (Make will already
remove any white-space character). So this check was removed.

This bug was reported by Johannes Zabl.