Corrected type of circumwidth in MakeProfiles
The type of `--circumwidth' in the options of MakeProfiles was defined to
be `float', but in `main.h', it was `double'. Therefore the value was
effectively being ignored (always read as zero, since only the first 4
bytes were written and the endian-ness of my CPU). The type is now
corrected and also the check is being made better to ignore pixels that are
farther out than the truncation radius (can happen right on the edge and
can be important for very small profiles).