General function to read arrays of any type
For tables, we have the general `gal_table_read' function which makes it
very easy to read the columns in the given file with any format. However,
until now we didn't have such a generic function for arrays. Such a
function will make it easy for all programs and library users to easily
read the input from all known types. This function has now been added to
the library.

During the process, `gal_fits_img_read' was also modified to not read the
WCS structure of the input FITS file. Because it would need two extra
arguments and the WCS is not always necessary, so there is no need to waste
time in such cases (reading and interpretting the WCS).

This was done as part of task #14885 (not yet complete).