Library is now in charge of writing EPS and PDF files
Until now, only the ConvertType program had the capabilities to write EPS
and PDF files. The respective functions are now brought out of ConvertType
and moved to the libraries to be accessible by all progarms and library
users.

Since the number of sections on the different (non-FITS) file formats in
the Gnuastro library chapter was getting too much, all the non-FITS formats
are now moved within a single section.