MakeCatalog can take a variance image instead of standard deviation
Some surveys provide the weight image in units of variance, not standard
deviation. So until now the user had to run Arithmetic to take its square
root and convert it to standard deviation, then feed the image into
MakeCatalog. Since this step is so trivial and an extra file can be really
annoying (especially on large files), a new option called `--variance' is
now available in MakeCatalog. When it is called, MakeCatalog will take the
square root of the input STD image prior to processing it.

Also, the NEWS file format was updated to a more clear to find format with
only one line per option/library/bug in the release.