Table's --column option can now take multiple columns
Until now, when using Table, it was necessary to identify each column with
a separate call to `--column' (or `-c'). This is annoying! It is now
possible to give multiple column identifiers (separated by a comma) to a
single call of this option. Ofcourse, it is still also possible to use
multiple calls to `column' or mixed: such that each call possibly has
multiple columns.