MakeCatalog's error is per-pixel, not average of all area
Until now, MakeCatalog's error was calculated like this: the average error
value over the object was found, then using the area and total flux, it was
converted to the total object's error that would be used in the
signal-to-noise estimation. This is good when the error (exposure time)
under the object is flat. However, on surveys, we don't usually have this
situation and the error can vary strongly on some parts (over one object).

So we now estimate the error on a per-pixel basis and add the errors up (as
variance). This made the signal-to-noise estimate much more
cleaner/readable and we MakeCatalog also now accepts a `--brightnesserr'
option to print the error in measuring the brightness.

While checking, I also noticed that a `break' was missing in the final
writing of the column values in MakeCatalog's `columns.c'. This is now
corrected.