New sigma-clipped mean and median filters for Arithmetic
The Arithmetic program now has two new filters: `filter-sigclip-mean' and
`filter-sigclip-median'. They can be used to remove strong outliers in the
filter before estimating the mean or median.

Arithmetic's filtering operators were also measuring the filter on blank
elements (in the input). This is not desired: a blank element in the input
should also be blank in the output. So a check was added to each pixel
before the filter is measured for it.

While trying to ignore blank pixels, I noticed that we don't have a simple
function to check a single value to see if its blank or not. So a new
function was defined for this job: `gal_blank_is'.

In the process, I also noticed that when using `gal_blank_remove' or
`gal_statistics_no_blank_sorted', we weren't checking for the case that all
the input pixels are blank. All occurances of these two functions now check
for this.