New Segment, NoiseChisel only detection, MakeCatalog re-write
Until now, NoiseChisel was in charge of both detection and
segmentation. This was making it extremely large and hard to manage, both
for the users (in terms of effectively understanding/using it) and the
developers (making it hard to add new features/options). Therefore
NoiseChisel's segmentation part has been spinned off into a new program
called Segment. It already has some good features (like giving an S/N of
true clumps to avoid having to find the S/N from the Sky regions)

As a result of this change in NoiseChisel and the new Segment program,
MakeCatalog (which uses the inputs of these programs) also underwent some
major changes in line with some several old tasks that were defined
before. In particular, it is now much more memory and CPU efficient: it
will only read the datasets that the output columns need and (when
necessary), it can read the Sky and Sky Standard deviation images from a
tiled image (with one pixel per tile).

This has been a very major commit and will hopefully open the way for many
more exciting features in NoiseChisel, Segment and MakeCatalog.

The General Program Usage Tutorial has not yet been corrected to
incorporate the new changes. It will be corrected shortly.

This completes tasks #13759, #14164, and #14869.