Arithmetic only reads the headers when necessary
Some file formats (like JPEG) don't need any headers. Therefore Arithmetic
will not attempt to assign them a HDU and will instead pass a NULL pointer
to their HDU element (as it should be).

Also, I found out that the `addcounter' element in Arithmetic's main
structure was extra and no longer used. So it was removed.