Warp checking for WCS when writing output
Until now, when writing its output, warp was implicitly assuming that the
dataset always has a WCS structure. Therefore when there was no WCS
structure, Warp would given a segmentation fault. To fix this, the
`correct_wcs_save_ouptut' function now checks if a WCS is present and only
does WCS operations when it is.

This bug was found thanks to a dataset provided by Bertrand Pain.

This fixes bug #53580.