Function to label holes in a binary image
This commit started with a testing a possible new strategy to fill holes in
NoiseChisel (during the growth of detections, using their S/N instead of
area). It was thus necessary to label the holes and
`gal_binary_holes_label' was written.

During the test, I also noticed that the `gal_data_ptr_dist' function name
was confusing (the "pointer-distance" count be interpretted as the
byte-base difference). The function is now named `gal_data_num_between' to
be more clear that it returns the number of elements between the pointers
in the type of the dataset.

Ultimate, the test didn't give good results: the S/N in the holes of the
diffuse regions have too much scatter to be meaningful. So the minor
function name change and new function will remain in the library, but
NoiseChisel still works as it did before.