Segment Sky and its STD stored in datasets
Until now if the input Sky or its standard deviation were a single value,
it was stored in `p->skyval' or `p->stdval' (both `float's). When it was a
dataset, it would be stored in datasets. However the two `float' elements
of the main structure were redundant (a dataset can also contain a single
element).

Therefore, the two `p->skyval' and `p->stdval' elements have been removed
and a dataset was used to keep both the sky value and its standard
deviation. To check if its a value or dataset, a simple test of their
`size' elements is now used.

This was previously employed in MakeCatalog. But I had forgot to bring it
here in Segment.