Segment can now output only clumps and avoid objects
In some contexts, an object image is not necessary so it is a waste of
resources (both CPU, memory and storage space) to force the user to keep an
objects labeled image also. Therefore a new `--onlyclumps' option was added
to Segment that will only output one labeled dataset and all the clumps
will have a unique label.