Updated NEWS file, minor correction in book's code example
In the previous commit, the NEWS file wasn't updated, so the new `erode'
and `dilate' operators of Arithmetic are now mentioned there too. Their
description was also slightly modified to be more clear.

Also, while flipping through the book, I noticed that one of the code
example lines in the library demos was a little too long (and pushing out
of the page). This has also been corrected.