Arithmetic's filtering operators also work on blanks
Until now, Arithmetic's filtering operations ignored blank valued
pixels. However, in some cases, the user might want to use the filtering to
remove/interpolate over the blank pixels. So the condition to ignore blank
pixels is now removed. A paragraph was added to the `filter-mean' operator,
describing the effect and how to correct for it if it is important to
preserve blank elements.