gal_box_bound_ellipse return width isn't extra
Until now, the width returned by `gal_box_bound_ellipse' would be 2 pixels
extra along each dimension, thus leaving some extra pixels on the edges
when MakeProfiles used it. This was done to accommodate extreme cases,
where the center of a profile was just on the edge of the pixel and the
trunctation radius also had a fractional value.

However, after some checks, we see that even in the most extreme cases,
only a single (very distant) elliptical layer may be lost (not at all
significant). But in most (normal cases), these extra layers will cause
many extra pixels which can be a burden on MakeProfiles its self or
higher-level programs (for example when building a kernel to convolve
with).

Also, while checking the `gal_box_bound_ellipse' functions, I noticed that
in the documentation, we were mistakenly describing `a' and `b' as
major/minor axis length. While in practice they are `semi-major' and
`semi-minor'.

Since there is no extra layer, we also don't need the call to Crop when
making NoiseChisel's default internal kernel (in
`lib/gnuastro-internal/kernel-2d.h'). A new kernel is now created and used
(with a larger number of random points).