Corrected range of --snquant in NoiseChisel and Segment
Until now, in Segment, the acceptable range for `--snquant' was only `>=0'
(or `GAL_OPTIONS_RANGE_GT_0'). However, a quantile can't be larger than
one, and it may be zero (the minimum). So it is now corrected to
`GAL_OPTIONS_RANGE_GE_0_LE_1'. In NoiseChisel, it was `>0 && <1' (or
`GAL_OPTIONS_RANGE_GT_0_LT_1'), so NoiseChisel's acceptable range for this
option was also corrected to `GAL_OPTIONS_RANGE_GE_0_LE_1'.