Segment's multiple check conditions now working properly
When Segment was called with the `--checksn' option, it would stop
producing check images (necessary for the other checking steps). This was
because of some small mistakes in the logics of the checks. With this
commit, these logical comparisons have been corrected and hopefully there
shouldn't be any more problem.