Segment's default minskyfrac decreased to 0.6 from 0.7
The 0.7 value for the `--minskyfrac' was slightly too much for many cases
(especially as the field get more crowded). So to be more useful in a
generic case, it was decreased to 0.6.

Also, to help in reading, the `arr' variable of `clumps_get_raw_info' was
changed to `values' and some comments were corrected.