Reading special floating point values in FITS ASCII tables
Until now, we would rely on CFITSIO to interpret the numerical values of
FITS ASCII tables (as we do for the binary tables). However, there was a
problem regarding special values like `INF' and `-INF'. In FITS ASCII
tables, CFITSIO can't interpret these values (that it writes itself).

To fix the problem, `fits_tab_read_ascii_float_special' was added. This
function will read the column contents as strings, then it will parse them
as floating point numbers with the robust `strtod' implementation of the
GNU C Library (which is shipped with Gnuastro for internal usage).

Also, the example code in `gal_fits_key_read_from_ptr' was corrected for a
few mistakes.