gal_label_oversegment now checks for blank values
Until now, `gal_label_oversegment' wouldn't check for blank values,
therefore, if the user hadn't already set the blank value flag, it wouldn't
put rivers on pixels touching blank values. So a call to
`gal_blank_present' was added to this function and a single variable is now
set to see if blank values are present or not (making it easier to read
than bit-flag checking deep into the loop).

Also, the description in the bit flag `GAL_DATA_FLAG_BLANK_CH' was edited
to be more clear.