Vatican Library's usage of FITS added in intro of Fits
Recently I learnt that due to the robustness and archivability features of
FITS, the Vatican Library (one of the oldest existing archival features in
the western world) is using the FITS format for its long term archives. So
to help encourage readers to use this format and get more familiar with it,
this point is added in the introduction to the Fits program.