New --ignoreblankinsky for NoiseChisel and Statistics
Until now, NoiseChisel and Statistics would set all the pixels that are
blank in the input to blank in the Sky and Sky standard deviation images
also. This was good for visual inspection when the field isn't aligned with
the rectangular image.

But in some situations (for example when you have intentionally set some
pixels to blank temporarily), you actually want the Sky and its standard
deviation under the blank pixels also. With this new `--ignoreblankinsky'
option, both NoiseChisel and Statistics will ignore the input's blank
values when writing the output with the same size as the input
(`--oneelempertile' wasn't called).