gal_qsort_index_arr initialized to NULL
We had defined `gal_qsort_index_arr' in `lib/qsort.c', but hadn't
initialized it. To be clean (definitely force a segmentation fault when the
user doesn't set it), it is now initialized to NULL. It is written to be
initialized to NULL in the comment, but apparently the actual initalization
was forgotten.