Sorting of indexs of multiple arrays in multiple threads
Until now we only had the `gal_qsort_index_float_increasing' function that
would use a global variable to keep the array to sort the indexs. This was
good when were were only dealing with a single array (and possibly many
arrays of indexs, on multiple threads). However, when each thread needs to
sort indexs that are defined in a separate array, that function would fail
to give a reasonable result.

With this commit, we now have the new `gal_qsort_index_multi_d' and
`gal_qsort_index_multi_i' functions for such situations. For such
situations, it is necessary to keep a pointer to the values array beside
each index. A structure (`gal_qsort_index_multi') is used to do this.

Also, to follow a logical naming convension, the old
`gal_qsort_index_float_increasing' and `gal_qsort_index_float_decreasing'
functions have been renamed to `gal_qsort_index_single_d' and
`gal_qsort_index_single_i'.

The whole purpose for this was that I wanted to run `gal_label_oversegment'
on separate datasets, each on a separate thread. Therefore,
`gal_label_oversegment' will now look into the flags of the `indexs'
array. It will only attempt to sort the indexs if they are not sorted. In
the multi-array and multi-threaded scenario above, indexs can be sorted
outside `gal_label_oversegment' (with the new `gal_qsort_index_single_d')
and the `indexs->flag' can be set. Afterwards `gal_label_oversegment' will
work cleanly on every thread/array separately.