NoiseChisel warning when the edges have a value of zero
Most data reduction pipelines put a value of 0.0 where there isn't any data
(commonly on the outer-frames of the observed field). But for NoiseChisel,
a value of 0.0 is meaningful, so it will give bad results. To help early
users understand this important point, a small check was added to see if
the first (and last) two pixels are 0.0 or not. If they are, a complete
warning message will be printed at the start of NoiseChisel's run and even
propose an solution (using Arithmetic).