NoiseChisel ignoring tiles with zero standard deviation
It might happen that the input dataset has zero values for regions with no
data (essentially using zero as a mask value). In other detection methods
(that have above-Sky thresholds) this isn't an issue. But NoiseChisel's
threshold is below the Sky value, so the zero valued pixels will actually
be considered as detections. We therefore encourge users to set zero valued
pixels to NaN/blank.

When there are tiles with fully zero-valued pixels, they may thus be used
for estimating the pseudo-detection Sky and Sky STD and result in
zero-valued Sky (and its STD) values. This will later cause the
counts-per-second correction (`cpscorr') to be zero and produce `inf' S/N
values.

With this commit, a check is added after finding the Sky and its Standard
deviation on a dataset. If the Standard deviation is zero, the tile will be
given a blank value (to be interpolated later).

This issue was found based on a dataset provided by Brandon Kelly.

This fixes bug #53825.