New tutorial on detecting large and extended targets
One of the most interesting science cases for low surface brightness
science is the large and extended objects (which usually cover large areas
of the image). Therefore estimating the noise properties for them is
particularly tricky. A tutorial was added to explain in detail the
thought-process of how the settings should be tuned to accurately detect
large extended objects to very low surface brightness limits (S/N of 0.05
for the example used in the tutorial).