Minor edites in new large object detection tutorial
Going over the large object detection tutorial one more time, some minor
edits were made to make it easier to read, fix typos and etc. Also a notice
was added to remind readers that this configuration is not generic for any
large object, and such cases are best analyzed when they are processed
individually/manually.