New true clump identification method
Until now, we were using the total clump signal-to-noise ratio (and its
distrubtion in the undetected regions), as a way to identify clumps. But
this didn't work well on gradients. So after several trials, the difference
between the maximum clump and river values was used instead. It apparently
works well on gradients and also more flatter regions.

Also in Segment, the default kernel was sharpened to a FWHM of 1.5
pixels. Therefore, the `kernel-2d.h' file is now separate between
NoiseChisel and Segment (each have their own copy).

Other minor corrections:

 - To fix a GCC 8.1 warning, the `basekeyname' statically allocated array
   is now 5 characters less than the full length.

 - `gal_statistics_quantile' can now account properly for an already sorted
   increasing or decreasing array. The quantile is defined on an increasing
   array, so if the array is already sorted in decreasing order, then we'll
   have to ask for the index of 1-quantile.

 - As suggested by Nushkia Chamba, the tutorial on large object detection
   was extended to include segmentation and identifying the background
   objects.