Absolute build and source directories for LDFLAGS and CPPFLAGS
It was a long time I hadn't ran `make distcheck'. But after doing so, I
noticed a libtool error, complaining about not finding the `./lib'
directory while it was building Gnulib (in `bootstrapped/lib'). After some
digging up in the code, I found the cause of the problem: at the start of
`configure.ac', we were adding `top_srcdir' and `top_builddir' (which are
relative paths) to `CPPFLAGS' and `LDFLAGS'. To fix the problem, we are now
using the absolute Autoconf paths: `abs_top_srcdir' and `abs_top_builddir'.