Clump significance measure now in library
After adding oversegmentation to the library, it was natural to also add
the function to measure the significance criteria on all the clumps. So the
old `clumps_make_sn_table' in `bin/segment/clumps.c' has been put in
`lib/label.c' with the name `gal_label_clump_significance'. It is now also
fully documented.