gal_label_watershed is the new name for gal_label_oversegment
The name "oversegment" is very old in Gnuastro. In those days, we called it
over-segmentation because it was building the profiles from the local
maximums, not local minimums as in the watershed algorithm. But now, it can
also build from the minimums. Since the watershed algorithm name is very
well known, it is thus better to use that name. Until version 0.5, this was
only available in the Segment program, so it wasn't visible to the outside
world and I had more important things to do. But now that it is indepedent
library function, its important to give it a recognizable name.