tests/during-dev.sh allows checks on BuildProgram natively
Until now, the `tests/during-dev.sh' script would always assume that the
top directory containing the program's configuration file is the source
directory. However, some programs (currently only BuildProgram) build their
configuration file during Gnuastro's build. With this commit, an `if'
statement has been added to avoid having to manually edit those parts of
this script any time we are testing such programs.

In the same `if' statement, we now also add the extra `--la' option, so
when testing with BuildProgram, it looks into the library built here, not
the one installed on the system.