Description of new true clump critera corrected in book
In the last few commits, a small modification was made to the criteria to
choose true clumps: the noise is now the actual noise taken under the river
pixel used (with maximum value). It is not calculated from the clump's
flux-weighted central position any more. This is more realistic (especially
now that we are only subtracting two pixel values).