Minor modifications for Debian release
While packaging for Debian, I noticed that for the experimental release, we
don't have a detached signature file. So the necessary step was added in
the checklist. This helps in decreasing the number of Lintian warnings,
allowing more cleaner output to focus on other more important issues.

Also, while building in Debian, I noticed a compiler warning regarding an
uninitialized variable in `tiff_read_separate_strip_data'. It was corrected
by initializing to zero.

Finally, a comment in `label_clump_significance_raw' was corrected.