Package manager commands for easy dependency installation
To make it easier to install Gnuastro's dependencies, the commands
necessary to install all the dependencies using the OS's package manager
have been added in the top "Dependencies" section. Also the discussion on
why its preferred to install packages from source (which was in the
"Mandatory dependencies" section) has been moved to the top "Dependencies"
section.