Initialize the first value in gal_label_clump_significance
Until now, when the `keepsmall' argument of `gal_label_clump_significance'
was non-zero, the first value (label/index 0) of the significance array was
un-initialized. This can cause confusion if the inspection of the results
isn't done from index one, but zero. To avoid un-pleasent surprizes for
early users of this function, when `keepsmall' is non-zero, it now sets the
`sigarr[0]' element to NaN.