AM_CPPFLAGS and AM_LDFLAGS contain internal libraries
To tell the programs where to find the Gnulib and Gnuastro library headers
and binary files, until now, we were modifying the global CPPFLAGS and
LDFLAGS in `configure.ac'. However, this caused several complications: 1)
forcing us to define `IN_CPPFLAGS' and `IN_LDFLAGS' for BuildProgram, 2)
forcing us to remove the extra addresses from the installed
`libgnuastro.la', 3) redundant (possible buggy) options when building and
linking Gnulib and the library.

To solve this problem, we are only adding the necessary directories in each
program's `Makefile.am' through the `AM_CPPFLAGS' and `AM_LDFLAGS' options.