Edits in the Synchronization section
Based on a discusstion with Leindert Boogaard, the Synchronization section
was edited to be more clear. Also, a `make distclean' step was added and we
now recommend `autoreconf' instead of `autoconf'. Also, a link to the
"Configure and build in RAM" section was added to introduce the readers to
the very convenient `tmpfs-config-make' script.