Options for the separate build directory script
Until now, the main tool for easy development was the `tmpfs-config-make'
script. But it didn't accept options and everytime we wanted to debug or
make some other modification to it, it was necessary to manually update it.

With this commit, it is renamed to "developer-build" and a whole section is
now devoted to it in the book, describing its options and how to run it. In
effect it has been separated from the "Configure and build in RAM" section.