New Dependencies from package managers section in doc
The start of the "Dependencies" section was getting too long. Also, it
would help to have "package managers" as a section title (to be included in
the contents). Therefore the discussion on package managers and the example
commands have been moved to a new subsection under "Dependencies", that is
called "Dependencies from package managers".

Some minor corrections were also made in the description of this
section. In particular, since Arch Linux is less common than Homebrew, its
description was moved below Homebrew. Also, the description of Homebrew was
broken up into two paragraphs for easier reading along with some small
edits.