tests/during-dev.sh internally adds -I and -L for BuildProgram
When testing BuildProgram with `tests/during-dev.sh', we had to manually
add the `-I' and `-L' options so it would look into the built Gnuastro, not
the one on the user's system. Previously we added program-specific options
to this script, so with this commit, we are using the same variable to add
`-I$srcdir/lib -L$builddir/lib'.