Wrappers for GSL's 1D interpolation added to library
Two basic wrappers were added to make it easy for using GSL's 1D
interpolation functions with datasets stored in Gnuastro's `gal_data_t'.

Some minor corrections in the indexs of the manual were also found and are
also commited here.

Also, while working on text files, I noticed that the error message when
trying to read a text table wasn't too clear, so it is made more clear with
this commit.