Corrected comments and description of developer-build
The comments at the start of the `developer-build' script still pointed the
reader to the old "Configure and build in RAM" section of the book (which
doesn't exist any more). So it was corrected to suggest running with
`--help'. Also, in the `--help' output, a new paragraph was added to
describe the other options and also refer the reader to the proper section
of the book.