gal_data_alloc defined before gal_data_initialize in code and book
`gal_data_alloc' is a higher-level function and more commonly used compared
to `gal_data_initialize'. To get to it in the manual however, first we have
to skim over `gal_data_initialize' (which is long). With this commit, we
are brigning the definition of `gal_data_alloc' (in the code and book,
which should have the same order) before `gal_data_initialize'.