Minor corrections in book and release checklist
After the release, I noticed that the first accent on Éric Thiébaut's name
was mistakenly placed on the `r' instead of the `E' in the book. It is
corrected now.

Also, a typo was found in the release checklist and Ole's email address was
added (to inform for the upload to Debian's servers).