GSL's Steffen interpolation checked at configure time
After the 0.6 release, Alan Lefor and Takashi Ichikawa informed us that
Gnuastro's build would fail on systems with an older GSL, because the
compiler couldn't find `gsl_interp_steffen'.

Therefore with this commit, a check has been added to the configure script.
If this variable isn't present in GSL's `gsl/gsl_interp.h' header, then
that type of interpolation will not be available and the program trying to
use it will crash with a notice. Also, at the end of the `./configure'
script, a warning will be printed, informing the user of this limited
capability of Gnuastro due to the old GSL version.

This fixes bug #54057.