Removed extra output option and log in tests/match/merged-cols.sh
The command to test Match's output when mixing the columns of two catalogs
had one call to `--output' and another to `-o'! The first one had remained
from the `tests/match/positions.sh' script (that was used as a basis for
this one): we had forgotten to correct it.

Also, the fact that both calls to Match in the tests used the `--log'
option caused a crash in Debian's build of Gnuastro 0.6. I had confronted
this once before on my system, but since I couldn't reproduce it and it
happened rarely (randomly, only once so far), I had forgot about it.

The problem is that the log file has a single name for any call. So when
the first call creates it and wants to write to it, but another call then
deletes it for its own use, we get this crash. So two calls to a progarm
with `--log' shouldn't be made in the same directory. A notice about this
important issue is now added in the description of the general `--log'
option and also in Invoking Match.

This fixes bug #54063.